requirements.txt
Django==5.0.2
djangorestframework==3.14.0
channels==4.0.0
django-cors-headers==4.3.1
Pillow==10.2.0
django-filter==23.5
django-cleanup==8.1.0
python-dotenv==1.0.1
django-storages==1.14.2
boto3==1.34.34
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
django-allauth==0.61.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-debug-toolbar==4.3.0
gunicorn==21.2.0
whitenoise==6.6.0
channels_redis==4.1.0
daphne==4.0.0
websockets==12.0 